split game done; make `HeroController` an interface
