+++ core/src/main/java/work/chiro/game/application/HeroController.java
@@ -1,0 +1,5 @@
+package work.chiro.game.application;
+
+public interface HeroController {
+    boolean isShootPressed();
+}
